pyproject.toml
[tool.poetry]
name = "drf-ecommerce"
version = "0.1.0"
description = ""
authors = ["João Feitoza <joaofeitoza.13@gmail.com>"]
maintainers = ["João Feitoza <joaofeitoza.13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
django-cors-headers = "^4.2.0"
django-mptt = "^0.14.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pylint-django = "^2.5.3"
python-dotenv = "^1.0.0"
pytest = "^7.4.1"
pytest-django = "^4.5.2"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
drf-spectacular = "^0.26.4"
coverage = "^7.3.1"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.5.1"
sqlparse = "^0.4.4"
pygments = "^2.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 95
target-version = ['py39']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 95
exclude = [".git", "__pycache__", "dist"]

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 95

[tool.djlint]
profile = "django"
ignore = "H031"
